major fixing of the learning mechanism
All previous versions of the code implemented the Rescorla-Wagner
learning rule in the wrong way; this problem has now been fixed.
With respect to the CogSci2016 submission, the results and their
interpretation don’t change, but the figures that are obtained change
slightly. The paper shows results from older commits, use this one to
generate the correct weights for the discriminative learning model.